updating docs workflow for oras-py to be updated nightly
It does not look like we are going to get a bot (and access token)
for the dispatch event anytime soon and I still want to make progress
on getting the docs where they need to be, so I am adding a nightly
dispatch that will currently open a PR, and when that works we can
change to just push. If there are no changes it will not do anything,
which is likely the most common case.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>